requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-MySQL==1.5.2
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
PyMySQL==1.1.1
Werkzeug==3.0.3
pip freeze > requirements.txt
Flask==2.0.1
gunicorn==20.1.0
